pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "privacyguard"
version = "1.0.0"
description = "Local Linux desktop text redaction tool for PII and secrets"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "PrivacyGuard Contributors"}
]
dependencies = [
  "packaging>=24.0",
  "platformdirs>=4.0",
  "requests>=2.31",
  "pystray>=0.19",
  "pillow>=10.0"
]

[project.optional-dependencies]
gui = [
  "PyGObject>=3.46"
]
nlp = [
  "spacy>=3.7"
]
secrets = [
  "detect-secrets>=1.5"
]
full = [
  "PyGObject>=3.46",
  "spacy>=3.7",
  "detect-secrets>=1.5"
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4"
]

[project.scripts]
privacyguard = "privacyguard.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["privacyguard*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
